Add initial navigation between list and detail views
Using the Navigation component, connect the two, including navigating back with the back button. 

There is no data being passed from the list to the details for now, so the details are still showing the hardcoded post data. Hitting the back button also currently causes the list view to completely reload. These should be fixed in a future change.