Preserve post list scroll state after viewing details and going back
When I was clicking on a post to view the details screen and then I clicked back, it would always reset the list view to the top of the screen. Reusing the same adapter instance by moving its creation to onCreate solves this case. However, if I rotate the device, it still resets the state, which isn't ideal. I have reason to believe it is possible to save this state in a sane way, so I will continue looking