Load details view post data from ID passed from list view
Use Navigation editor arguments to pass the ID of the post (which is now added to the data model) from one screen to another. Then, use this ID to pull the cached post from the repository (every time we retrieve posts, we add them to the cache).

The reason to use the ID passed between views is that in the future, I think I'd like to use deep linking to open the details view from a URL, so parameterizing the view with the ID will let me do that more easily.

There is probably a better way to handle the cache, but it is fine for now.

Also, slight cleanup of the map in the repository - use Kotlin's built-in collection mapping to handle things more compactly/idiomatically