Add flag to determine if  post is an external link
Use the isSelf flag. This is a rough approximation if the post is an external link or not. There are cases where it isn't actually an external link (for example, a link to another subreddit), but it's close enough, and should still work in those cases anyway (even if not in the most ideal fashion)