Break out usage of API vs app model for posts
This will officially switch the rest of the app away from using the API model, and conversion from one to the other will occur in the repository (for now... it may make more sense to have it actually in the network layer, but this is simple enough to implement for now).

This also renames a few models to make it clear they are for the API as well.