Update labels for nav destinations
Since the AppBar is now connected to the navigation controller, it will show the destination label in the AppBar. These were originally just default values, but they weren't really appropriate for end users. Remove the label for the post listing so it just shows the app name in the title bar, and add a "good enough for now" label for post details. eventually, I might want to make these something different (like the name of the subreddit), but this is at least better that the default labels.