Add todo for thing that is broken
The way I had this before, it would cache the results of the network call, which meant that it wouldn't re-fetch the data on subsequent calls. Since this got changed though, it redoes the call because this is now a cold observable again. I think Switching to LiveData would be a good fix for this though.