Experiment with Groupie for comment section
The Groupie library seems like it has good potential for nested comment stuff, so play around with it for a bit.

In the process, I found this article that is basically doing the same thing, so I was using it for a bit of reference: https://medium.com/@pedrocarrillo/comments-in-a-recylcerview-e7537235cd33

I might put this on hold for a bit though since it seems like it could be complicated, and there are other things I wanna do first.