Add extension for displaying large numbers
Basically, if a number is greater than 1000, show the value in thousands (with a "K" at the end) and one decimal place.

I'm not entirely sure I love packaging this extension in a thing called UiExtensions, but I can't really think of a better place off the top of my head.