Avoid refetching post list on fragment recreation
I had this working at one point before, but it got broken at some point. Use LiveData to propagate the post list to the fragment. This will allow us to make the Rx call once in the ViewModel, and then using the caching mechanism of LiveData to show the same data again upon recreation. This avoids churn if the user rotates the screen and causes potentially different data to load.