Add tests for display number formatting
Note that this actually helped identify a bug in the code around the case where the number was exactly 1000. However, I left this intentionally broken, but with the test expecting the actual desired value - meaning that the test fails right now. This is intentional because I'm also trying out a CI solution, and I would like to see what a failure looks like.