Add "simple" test for RedditPostRepository
This basically tests a simple, happy-path case where the response contains a single item.

This could be cleaned up a fair amount, including extracting the test and expected model objects into separate methods, but it's good enough or now since this is the only test.

It also doesn't appropriately test that the stream operates on an IO scheduler, but this can be done later.