Add title for Image page
This fixes an issue on the image screen where the screen title resets to app name when rotating the screen.

I suspect the underlying issue is a bug either in the navigation component or in the general app bar code. I should probably file a bug for it, but just haven't gotten around to it.

I don't really like the "Image" text here because it seems too generic, but it's at least better than the title acting weird.