Add fullname field to reddit listing/post
As I came to learn, this is needed for determining which "page" you're on when making requests for more posts.

This field is essentially the post type (like "t3") followed by the id with an underscore in-between. I could probably optimize this a bit by reconstructing the field value from those fields, but it seems like a lot of work, and it's already provided by the API anyway, so just use it out of convenience.