Add ViewModel method for fetching more posts
To achieve this, the ViewModel holds a cached version of the full post list. I'm ultimately not sure if this makes more sense to do in the repository somehow, but I ended up going this route because this helped with resetting the cached list if the user navigated out of the scope of the ViewModel.

Like in the previous commit, there is some duplicated logic which should be cleaned up eventually, but I'm more interested in getting everything working first and then cleaning up later - hence the FIXME comment.