Fetch more posts if user is near end of the list
This code detects if the user is close to scrolling to the end of the post list, and if so tells the ViewModel to fetch more.

There is some commentary in here wondering if this is the best solution or not. For now, it works, so we'll go with it.

Note that I also removed the layout manager tag from the XML and instantiated it in the fragment instead. While this is generally not my preference, the fragment was relying on the use of a LinearLayoutManager anyway to detect the last item, so I figured I might as well make that implementation detail encapsulated in the Fragment rather than having the Fragment coupled to details in the XML (which if changed could end up breaking things).