Add custom deserialization for ApiResponse
As I came to find out, replies to comments in the API are basically just complete responses of their own (as in, they map to the same data structure as the ApiRedditResponse). However, in at least some (if not all) cases where a comment doesn't have any replies, it would have an empty string value (instead of the field either being absent or having a null value). So, I basically have to add custom deserialization to handle this case and just return null when the object couldn't be otherwise deserialized.