Add class for fetching Comments
It's sort of like a repository, but I've started switching to calling these use cases for now. Anyway, it basically makes the network call to fetch these via the new method on the network service, and then converts the response into a list of Comment objects (note that this is the "app" model, not the API model anymore).

I did take the lazy route and made fields on the Comment nullable because they are technically nullable in the response for now. I should probably find a better way to handle this (perhaps by breaking the API models up so that I don't need to make fields nullable? See https://trello.com/c/mjsgTDIS).

Also, add new class to Koin for DI