Add comment expand touch target
For some reason that I don't quite understand, the constraint layout itself won't respond to clicks to expand replies, but the child views of the constrain layout would. So, I added another child view that represents the clickable area for expansion and set the click listener on that.

It unfortunately feels a bit hacky for some reason. On the other hand, it has the benefit of not letting users click on the depth indicators to expand/collapse replies, which seemed kind of weird to do.